src/main/java/com/github/karixdev/youtubethumbnailranking/auth/AuthService.java
package com.github.karixdev.youtubethumbnailranking.auth;

import com.github.karixdev.youtubethumbnailranking.auth.payload.request.RegisterRequest;
import com.github.karixdev.youtubethumbnailranking.shared.payload.response.SuccessResponse;
import com.github.karixdev.youtubethumbnailranking.user.UserRole;
import com.github.karixdev.youtubethumbnailranking.user.UserService;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;

@Service
@RequiredArgsConstructor
public class AuthService {
    private final UserService userService;

    @Transactional
    public SuccessResponse registerNewUser(RegisterRequest payload) {
        userService.createUser(
                payload.getEmail(),
                payload.getUsername(),
                payload.getPassword(),
                UserRole.ROLE_USER
        );

        return new SuccessResponse();
    }

}
